Begin algorithm rewrite with better data structures
fdupes data structures were very simple and ill-suited for the
sort of work fdupes does. This is the beginning of a large-scale
rewrite of fdupes to use more suitable data structures that could
increase performance and (more importantly) open the way for
much more sophisticated features in the future.

Existing code did not perform sorting of files correctly and made
many performance optimizations more complex than they needed to
be. Of particular note is that all files were compared to each
other and quickly excluded as duplicates by their size, but since
this is by far the most common duplicate exclusion criterion,
walking the entire list of files for each candidates doesn't make
any sense. The new data structure for storing candidates is a
linked list of sizes, each element of which serves as the head of
a linked list of files that are of that particular size.

An examination of a compiled linux-4.1 kernel source directory
shows that out of 65211 files, 14370 (22%) are of a unique size
and can be excluded from consideration by fdupes entirely.
Furthermore, the most common count of non-unique file sizes was
2 files at 4370 (6.7%) and the largest count of non-unique file
sizes was 2449 files (3.8%). In fact, the vast majority of
duplicate candidates have less than 14 other candidates when
grouped by size alone.

These numbers (which I admittedly may not have stated very well)
indicate that the number of comparisons for any given set of
potential duplicates is often a very small count, and the worst
case requires comparing a list of files that is a small fraction
of what fdupes has been comparing. I added some counters to the
checkmatch() function to get actual "hard" numbers on this and
found that out of 1,617,397 calls to checkmatch(), only 84,548
were NOT excluded based on file size. That's 1,532,849 match
checks and function calls that were not necessary, accounting for
94.8% of all checkmatch() calls.